PCODAR: Rename c++ extensions - hh -> hpp, cc -> cpp
