+++ perception/point_cloud_object_detection_and_recognition/include/point_cloud_object_detection_and_recognition/point_cloud_builder.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "pcodar_params.hh"
-#include "pcodar_types.hh"
+#include "pcodar_params.hpp"
+#include "pcodar_types.hpp"
 
 #include <sensor_msgs/PointCloud2.h>
 #include <tf2/LinearMath/Transform.h>
