+++ perception/point_cloud_object_detection_and_recognition/include/point_cloud_object_detection_and_recognition/point_cloud_clusterer.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "pcodar_types.hh"
+#include "pcodar_types.hpp"
 
 #include <mil_msgs/PerceptionObject.h>
