+++ perception/point_cloud_object_detection_and_recognition/nodes/main.cpp
@@ -1,4 +1,4 @@
-#include <point_cloud_object_detection_and_recognition/pcodar_controller.hh>
+#include <point_cloud_object_detection_and_recognition/pcodar_controller.hpp>
 
 int main(int argc, char* argv[])
 {
