+++ perception/point_cloud_object_detection_and_recognition/tools/labeling_tool/main.cpp
@@ -1,8 +1,8 @@
-#include <point_cloud_object_detection_and_recognition/marker_manager.hh>
-#include <point_cloud_object_detection_and_recognition/pcodar_params.hh>
-#include <point_cloud_object_detection_and_recognition/point_cloud_builder.hh>
-#include <point_cloud_object_detection_and_recognition/point_cloud_clusterer.hh>
-#include <point_cloud_object_detection_and_recognition/pcodar_controller.hh>
+#include <point_cloud_object_detection_and_recognition/marker_manager.hpp>
+#include <point_cloud_object_detection_and_recognition/pcodar_params.hpp>
+#include <point_cloud_object_detection_and_recognition/point_cloud_builder.hpp>
+#include <point_cloud_object_detection_and_recognition/point_cloud_clusterer.hpp>
+#include <point_cloud_object_detection_and_recognition/pcodar_controller.hpp>
 
 #include <eigen_conversions/eigen_msg.h>
 
@@ -37,8 +37,8 @@
 #include <QDesktopWidget>
 #include <QStyle>
 
-#include "label_model.hh"
-#include "main_window.hh"
+#include "label_model.hpp"
+#include "main_window.hpp"
 
 /*
 UI Features I want:
