+++ perception/point_cloud_object_detection_and_recognition/tools/labeling_tool/main_window.hpp
@@ -3,7 +3,7 @@
 #include <QMainWindow>
 #include <QtWidgets>
 
-#include "label_model.hh"
+#include "label_model.hpp"
 
 #include <memory>
